requirements.txt
pdfminer.six
chardet